Ensure each value returned is a unique object
